auth_service/exceptions.py
from rest_framework.views import exception_handler
from rest_framework.response import Response
from rest_framework import status
import logging

logger = logging.getLogger(__name__)


def custom_exception_handler(exc, context):
    """
    Custom exception handler for DRF
    """
    response = exception_handler(exc, context)

    if response is not None:
        custom_response_data = {
            'success': False,
            'error': {
                'message': str(exc),
                'detail': response.data
            }
        }
        response.data = custom_response_data

    else:
        logger.error(f"Unhandled exception: {exc}", exc_info=True)
        custom_response_data = {
            'success': False,
            'error': {
                'message': 'An unexpected error occurred',
                'detail': str(exc)
            }
        }
        response = Response(custom_response_data, status=status.HTTP_500_INTERNAL_SERVER_ERROR)

    return response
