requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-ratelimit==4.1.0

# Firebase
firebase-admin==6.3.0

# Brevo (formerly Sendinblue)
sib-api-v3-sdk==7.6.0
urllib3<2.0  # Required for compatibility with sib-api-v3-sdk

# HubSpot CRM
requests==2.31.0

# Security
django-csp==3.8

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
python-dotenv==1.0.0
